+++ warp/config.py
@@ -7,7 +7,7 @@
 
 import os
 
-version = "0.1.24"
+version = "0.1.25"
 
 cuda_path = None        # path to local CUDA toolchain, if None at init time warp will attempt to find the SDK using CUDA_PATH env var
 host_path = None        # path to local x86 toolchain, if None at init time warp will attempt to find MSVC toolchain